@@ -5,24 +5,24 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   bloc: ^7.0.0
-  flutter_bloc: ^7.0.0
-  flutter_screenutil: ^4.0.2+3
-  cupertino_icons: ^1.0.2
-  shared_preferences: ^2.0.5
-  http: ^0.12.2
-  flutter_dotenv: ^3.1.0
-  fluttertoast: ^7.1.6
+  flutter_bloc: ^7.0.1
+  flutter_screenutil: ^5.0.0+2
+  cupertino_icons: ^1.0.3
+  shared_preferences: ^2.0.6
+  http: ^0.13.3
+  flutter_dotenv: ^5.0.0
+  fluttertoast: ^8.0.7
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_gen_runner:
+  flutter_gen_runner: ^3.1.1
 
 
 
