@@ -4,25 +4,29 @@ import 'package:flutter/material.dart';
 
 /// Custom shapes for easy implementation for rounded borders ecc..
 
-const _circularRadius = 24.0;
+class AppShapes{
 
-BorderRadius appBorderRadius = BorderRadius.circular(_circularRadius);
+  static const _circularRadius = 16.0;
 
-Decoration containerElevationShadow({Key key, bgColor}) {
-  return BoxDecoration(
-    color: bgColor,
-    borderRadius: BorderRadius.all(Radius.circular(_circularRadius)),
-    boxShadow: [
-      BoxShadow(color: Colors.black12, spreadRadius: 1, blurRadius: 10),
-    ],
+  static BorderRadius appBorderRadius = BorderRadius.circular(_circularRadius);
+
+  static Decoration containerElevationShadow({Key? key, bgColor}) {
+    return BoxDecoration(
+      color: bgColor,
+      borderRadius: BorderRadius.all(Radius.circular(_circularRadius)),
+      boxShadow: [
+        BoxShadow(color: Colors.black12, spreadRadius: 1, blurRadius: 10,),
+      ],
+
+    );
+  }
+
+  static RoundedRectangleBorder appCardShape = RoundedRectangleBorder(
+    borderRadius: appBorderRadius,
+  );
+
+  static OutlineInputBorder appInputBorder = OutlineInputBorder(
+    borderRadius: BorderRadius.circular(_circularRadius),
+    borderSide: BorderSide(color: Colors.black26, width: 1.0),
   );
 }
-
-RoundedRectangleBorder appCardShape = RoundedRectangleBorder(
-  borderRadius: appBorderRadius,
-);
-
-OutlineInputBorder appInputBorder = OutlineInputBorder(
-  borderRadius: BorderRadius.circular(_circularRadius),
-  borderSide: const BorderSide(color: Colors.grey, width: 1.0),
-);
