@@ -1,7 +1,7 @@
 part of 'theme_cubit.dart';
 
 class ThemeState {
-  ThemeState({@required this.themeMode});
+  ThemeState({required this.themeMode});
 
   final ThemeMode themeMode;
 
