@@ -11,56 +11,53 @@ part 'login_state.dart';
 
 enum LoginStatus { Authenticated, Unauthenticated, Uninitialized }
 
-loadView(context, route) async {
-  await Future.delayed(const Duration(seconds: 1), () {
-    Navigator.pushNamedAndRemoveUntil(
-        context, route, (Route<dynamic> route) => false);
-  });
-}
-
 class LoginCubit extends Cubit<LoginState> {
-  User user;
-  final AuthRepository repository;
-  LoginStatus status;
+  User? user;
+  LoginStatus? status;
 
-  LoginCubit(this.repository) : super(LoginState()) {
+  LoginCubit() : super(LoginState()) {
     loginOnStartUp();
   }
 
+  /// Event invoked on startup to check if the user has the data saved on the device
   Future<void> loginOnStartUp() async {
-    emit(LoginState(
-        loading: false, user: null, status: LoginStatus.Uninitialized));
-    user = await repository.getUser();
+    emit(LoginState(loading: false, user: null, status: LoginStatus.Uninitialized));
+    user = await AuthRepository.getUser();
     if (user != null) {
-      emit(LoginState(
-          user: user, status: LoginStatus.Authenticated, loading: false));
-    } else {
-      emit(LoginState(
-          user: null, status: LoginStatus.Unauthenticated, loading: false));
-    }
-  }
-
-  Future<void> login({Key key, String email, String password}) async {
-    emit(LoginState(user: user, status: status, loading: true));
-    ResponseMessage<User> response = await repository.authenticate(email: email, password: password);
-
-    if (response.body != null && response.status == ResponseMessageStatus.s200) {
-      user = response.body;
-      await repository.persistUser(user);
       emit(LoginState(user: user, status: LoginStatus.Authenticated, loading: false));
     } else {
-      showToastTop(message: 'Credenziali errate!', bgColor: Theme.of(navigatorKey.currentContext).backgroundColor);
       emit(LoginState(user: null, status: LoginStatus.Unauthenticated, loading: false));
     }
   }
 
-  Future<void> logout() async {
-    emit(LoginState(user: user, status: status, loading: true));
-    await repository.deleteUser();
-    emit(LoginState(
-        user: null, status: LoginStatus.Unauthenticated, loading: false));
+  /// Event to login
+  Future<void> login({Key? key, String? email, String? password}) async {
+    // start loading animation
+    emit(LoginState(loading: true, user: user, status: status));
+    // make the authentication request to the backend
+    ResponseMessage<User?> response = await AuthRepository.authenticate(email: email, password: password);
+    print(response);
+    //check the response status
+    if (response.status == ResponseMessageStatus.s200) {
+      user = response.body;
+      await AuthRepository.persistUser(user!);
+      // if successful login
+      emit(LoginState(user: user, status: LoginStatus.Authenticated, loading: false));
+    } else {
+      // error message
+      AppToasts.showToastBottom(message: response.message, context: navigatorKey.currentContext!);
+      emit(LoginState(user: null, status: LoginStatus.Unauthenticated, loading: false));
+    }
   }
 
+  /// Event to logout the user
+  Future<void> logout() async {
+    emit(LoginState(user: user, status: status, loading: true));
+    await AuthRepository.deleteUser();
+    emit(LoginState(user: null, status: LoginStatus.Unauthenticated, loading: false));
+  }
+
+  /// Event to check if the user is Authenticated
   bool isAuth() {
     if (user != null && status == LoginStatus.Authenticated) {
       return true;
