@@ -1,11 +1,11 @@
 part of 'login_cubit.dart';
 
 class LoginState {
-  LoginState({Key key, this.user,this.status,this.loading});
+  LoginState({Key? key, this.user,this.status,this.loading});
 
   final user;
-  final LoginStatus status;
-  final bool loading;
+  final LoginStatus? status;
+  final bool? loading;
 
   @override
   String toString() {
