@@ -1,14 +1,15 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class SplashScreenPage extends StatelessWidget {
-  const SplashScreenPage({Key key}) : super(key: key);
+  const SplashScreenPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: Center(
         child: Image.asset('assets/images/splash.png',
-          width: MediaQuery.of(context).size.width * 0.785,
+          width: 300.w,
         ),
       ),
     );
