@@ -1,5 +1,5 @@
-// Copyright (c) 2021, Very Good Ventures
-// https://verygood.ventures
+// Copyright (c) 2021, Calin Vasile Andrei
+// https://calinandrei.com
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file or at
