@@ -8,8 +8,8 @@ class AppTheme {
   static Color lightPrimaryColor = const Color(0xfff2f2f2);
   static Color lightAccentColor = Color(0xffFA7B58);
   static Color lightParticlesColor = const Color(0x44948282);
-  static const Color lightHeadlineColor = Colors.black54;
-  static const Color lightParagraphColor = Colors.black;
+  static const Color lightHeadlineColor = Colors.black45;
+  static const Color lightParagraphColor = Colors.black87;
 
   static Color darkBackgroundColor = const Color(0xFF1A2127);
   static Color darkPrimaryColor = Colors.white;
@@ -22,160 +22,186 @@ class AppTheme {
     headline1: TextStyle(
         color: darkHeadlineColor,
         fontFamily: 'Lato',
+        letterSpacing: -1.5,
         fontWeight: FontWeight.w900,
-        fontSize: 56.sp,
+        fontSize: 96.sp,
         fontStyle: FontStyle.normal),
     headline2: TextStyle(
         color: darkHeadlineColor,
+        letterSpacing: -0.5,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w900,
-        fontSize: 52.sp,
+        fontSize: 60.sp,
         fontStyle: FontStyle.italic),
     headline3: TextStyle(
         color: darkHeadlineColor,
+        letterSpacing: 0,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w700,
         fontSize: 48.sp,
         fontStyle: FontStyle.normal),
     headline4: TextStyle(
         color: darkHeadlineColor,
+        letterSpacing: 0.25,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w700,
-        fontSize: 48.sp,
+        fontSize: 34.sp,
         fontStyle: FontStyle.italic),
     headline5: TextStyle(
         color: darkHeadlineColor,
+        letterSpacing: 0,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w700,
-        fontSize: 46.sp,
+        fontSize: 24.sp,
         fontStyle: FontStyle.normal),
     headline6: TextStyle(
         color: darkHeadlineColor,
+        letterSpacing: 0.15,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w700,
-        fontSize: 44.sp,
+        fontSize: 20.sp,
         fontStyle: FontStyle.italic),
     subtitle1: TextStyle(
         color: darkParagraphColor,
+        letterSpacing: 0.15,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w400,
-        fontSize: 44.sp,
+        fontSize: 16.sp,
         fontStyle: FontStyle.normal),
     subtitle2: TextStyle(
         color: darkParagraphColor,
+        letterSpacing: 0.1,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w400,
-        fontSize: 40.sp,
+        fontSize: 14.sp,
         fontStyle: FontStyle.normal),
     bodyText1: TextStyle(
         color: darkParagraphColor,
+        letterSpacing: 0.5,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w400,
-        fontSize: 36.sp,
+        fontSize: 16.sp,
         fontStyle: FontStyle.normal),
     bodyText2: TextStyle(
         color: darkParagraphColor,
+        letterSpacing: 0.25,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w300,
-        fontSize: 32.sp,
+        fontSize: 14.sp,
         fontStyle: FontStyle.normal),
-    button: TextStyle(
+    button: TextStyle(//capitalized
         color: darkHeadlineColor,
+        letterSpacing: 1.25,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w400,
-        fontSize: 40.sp,
+        fontSize: 14.sp,
         fontStyle: FontStyle.normal),
     caption: TextStyle(
         color: darkParagraphColor,
+        letterSpacing: 0.4,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w300,
-        fontSize: 28.sp,
+        fontSize: 12.sp,
         fontStyle: FontStyle.normal),
-    overline: TextStyle(
+    overline: TextStyle( //capitalized
         color: darkParagraphColor,
+        letterSpacing: 1.5,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w100,
-        fontSize: 28.sp,
+        fontSize: 10.sp,
         fontStyle: FontStyle.italic),
   );
   static TextTheme lightTextTheme = TextTheme(
     headline1: TextStyle(
         color: lightHeadlineColor,
         fontFamily: 'Lato',
+        letterSpacing: -1.5,
         fontWeight: FontWeight.w900,
-        fontSize: 56.sp,
+        fontSize: 96.sp,
         fontStyle: FontStyle.normal),
     headline2: TextStyle(
         color: lightHeadlineColor,
+        letterSpacing: -0.5,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w900,
-        fontSize: 52.sp,
+        fontSize: 60.sp,
         fontStyle: FontStyle.italic),
     headline3: TextStyle(
         color: lightHeadlineColor,
+        letterSpacing: 0,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w700,
         fontSize: 48.sp,
         fontStyle: FontStyle.normal),
     headline4: TextStyle(
         color: lightHeadlineColor,
+        letterSpacing: 0.25,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w700,
-        fontSize: 48.sp,
+        fontSize: 34.sp,
         fontStyle: FontStyle.italic),
     headline5: TextStyle(
         color: lightHeadlineColor,
+        letterSpacing: 0,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w700,
-        fontSize: 46.sp,
+        fontSize: 24.sp,
         fontStyle: FontStyle.normal),
     headline6: TextStyle(
         color: lightHeadlineColor,
+        letterSpacing: 0.15,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w700,
-        fontSize: 44,
+        fontSize: 20.sp,
         fontStyle: FontStyle.italic),
     subtitle1: TextStyle(
         color: lightParagraphColor,
+        letterSpacing: 0.15,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w400,
-        fontSize: 44.sp,
+        fontSize: 16.sp,
         fontStyle: FontStyle.normal),
     subtitle2: TextStyle(
         color: lightParagraphColor,
+        letterSpacing: 0.1,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w400,
-        fontSize: 40.sp,
+        fontSize: 14.sp,
         fontStyle: FontStyle.normal),
     bodyText1: TextStyle(
         color: lightParagraphColor,
+        letterSpacing: 0.5,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w400,
-        fontSize: 36.sp,
+        fontSize: 16.sp,
         fontStyle: FontStyle.normal),
     bodyText2: TextStyle(
         color: lightParagraphColor,
+        letterSpacing: 0.25,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w300,
-        fontSize: 32.sp,
+        fontSize: 14.sp,
         fontStyle: FontStyle.normal),
-    button: TextStyle(
+    button: TextStyle(//capitalized
         color: lightHeadlineColor,
+        letterSpacing: 1.25,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w400,
-        fontSize: 40.sp,
+        fontSize: 14.sp,
         fontStyle: FontStyle.normal),
     caption: TextStyle(
         color: lightParagraphColor,
+        letterSpacing: 0.4,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w300,
-        fontSize: 28.sp,
+        fontSize: 12.sp,
         fontStyle: FontStyle.normal),
-    overline: TextStyle(
+    overline: TextStyle( //capitalized
         color: lightParagraphColor,
+        letterSpacing: 1.5,
         fontFamily: 'Lato',
         fontWeight: FontWeight.w100,
-        fontSize: 28.sp,
+        fontSize: 10.sp,
         fontStyle: FontStyle.italic),
   );
 
@@ -185,6 +211,7 @@ class AppTheme {
       brightness: Brightness.light,
       primaryColor: lightPrimaryColor,
       accentColor: lightAccentColor,
+      buttonColor: lightHeadlineColor,
       backgroundColor: lightBackgroundColor,
       visualDensity: VisualDensity.adaptivePlatformDensity,
       textTheme: lightTextTheme);
@@ -193,12 +220,13 @@ class AppTheme {
       brightness: Brightness.dark,
       primaryColor: darkPrimaryColor,
       accentColor: darkAccentColor,
+      buttonColor: darkHeadlineColor,
       backgroundColor: darkBackgroundColor,
       visualDensity: VisualDensity.adaptivePlatformDensity,
       textTheme: darkTextTheme);
 
   static Brightness get currentSystemBrightness =>
-      SchedulerBinding.instance.window.platformBrightness;
+      SchedulerBinding.instance!.window.platformBrightness;
 
   static void setStatusBarAndNavigationBarColors(ThemeMode themeMode) {
     SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
