@@ -1,5 +1,5 @@
-// Copyright (c) 2021, Very Good Ventures
-// https://verygood.ventures
+// Copyright (c) 2021, Calin Vasile Andrei
+// https://calinandrei.com
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file or at
@@ -12,11 +12,11 @@ import 'package:flutter/widgets.dart';
 import 'package:bloc/bloc.dart';
 import 'package:flutter_easy_start/app/app.dart';
 import 'package:flutter_easy_start/app/app_bloc_observer.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart' as DotEnv;
+import 'package:flutter_dotenv/flutter_dotenv.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await DotEnv.load();
+  await dotenv.load();
   Bloc.observer = AppBlocObserver();
   FlutterError.onError = (details) {
     log(details.exceptionAsString(), stackTrace: details.stack);
