@@ -21,67 +21,70 @@ class _LoginPageState extends State<LoginPage> {
         FocusScope.of(context).requestFocus(new FocusNode());
       },
       child: Scaffold(
-        resizeToAvoidBottomInset: false,
+        backgroundColor: Theme.of(context).backgroundColor,
         body: SafeArea(
           child: Container(
-            width: ScreenUtil().screenWidth,
-            height: ScreenUtil().screenHeight,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              mainAxisAlignment: MainAxisAlignment.start,
-              children: [
-                Container(
-                    margin: EdgeInsets.only(top: 100.h),
-                    width: 200,
-                    height: 200,
-                    child: new Image.asset('assets/images/splash.png')),
-                InputTextField(
-                    width: ScreenUtil().screenWidth * 0.8,
-                    height: 150.h,
-                    marginVertical: 30.h,
-                    controller: emailController,
-                    enabledTextField: true,
-                    icon: Icons.email,
-                    hintText: "Inserire l'email"),
-                InputTextField(
-                    width: ScreenUtil().screenWidth * 0.8,
-                    height: 150.h,
-                    obfuscaredText: true,
-                    marginVertical: 30.h,
-                    controller: passwordController,
-                    enabledTextField: true,
-                    icon: Icons.security,
-                    hintText: 'Inserire la password'),
-                Container(
+            width: MediaQuery.of(context).size.width,
+            child: Padding(
+              padding: EdgeInsets.symmetric(horizontal: 24.w),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.center,
+                mainAxisAlignment: MainAxisAlignment.start,
+                children: [
+                  Container(
+                      width: 200.w,
+                      height: 200.h,
+                      child: new Image.asset('assets/images/splash.png')),
+                  InputTextField(
+                      width: 320.w,
+                      height: 60.h,
+                      controller: emailController,
+                      enabledTextField: true,
+                      icon: Icons.email,
+                      hintText: "Inserire l'email"),
+                  SizedBox(height: 20.h),
+                  InputTextField(
+                      width: 320.w,
+                      height: 60.h,
+                      obfuscaredText: true,
+                      controller: passwordController,
+                      enabledTextField: true,
+                      icon: Icons.security,
+                      hintText: 'Inserire la password'),
+                  Container(
+                      width: 320.w,
+                      height: 60.h,
+                      margin: EdgeInsets.only(top: 30.h),
+                      child: ElevatedButton(
+                          onPressed: () {
+                            FocusScope.of(context).requestFocus(new FocusNode());
+                            BlocProvider.of<LoginCubit>(context).login(
+                                email: emailController.text.trim(),
+                                password: passwordController.text.trim());
+                          },
+                          style: ElevatedButton.styleFrom(
+                            primary: Theme.of(context).accentColor,
+                            shape: AppShapes.appCardShape,
+                          ),
+                          child: Text(
+                            'LOGIN',
+                            style: Theme.of(context).textTheme.button!.copyWith(
+                                color: Theme.of(context).backgroundColor),
+                          ))),
+                  Container(
                     margin: EdgeInsets.only(top: 50.h),
-                    width: ScreenUtil().screenWidth * 0.8,
-                    height: 150.h,
-                    child: ElevatedButton(
-                        onPressed: () {
-                          FocusScope.of(context).requestFocus(new FocusNode());
-                          BlocProvider.of<LoginCubit>(context).login(
-                              email: emailController.text.trim(),
-                              password: passwordController.text.trim());
-                        },
-                        style: ElevatedButton.styleFrom(
-                          primary: Theme.of(context).accentColor,
-                          shape: appCardShape,
-                        ),
-                        child: Text(
-                          'Login',
-                          style: Theme.of(context).textTheme.bodyText1.copyWith(
-                              color: Theme.of(context).backgroundColor),
-                        ))),
-                BlocBuilder<LoginCubit, LoginState>(builder: (context, state) {
-                  if (state.loading) {
-                    return Container(
-                        margin: EdgeInsets.only(top: 100.h),
-                        child: CircularProgressIndicator());
-                  } else {
-                    return Container();
-                  }
-                }),
-              ],
+                    width: 30.w,
+                    height: 30.h,
+                    child: BlocBuilder<LoginCubit, LoginState>(builder: (context, state) {
+                      if (state.loading!) {
+                        return CircularProgressIndicator();
+                      } else {
+                        return Container();
+                      }
+                    }),
+                  ),
+                ],
+              ),
             ),
           ),
         ),
